16-bit framebuffer looks better

+++ rcube/systems/RenderSystem.cpp
@@ -33,15 +33,15 @@ unsigned int RenderSystem::priority() const {
 
 void RenderSystem::initialize() {
     framebufferms_ = Framebuffer::create(resolution_[0], resolution_[1]);
-    framebufferms_->addColorAttachment(TextureInternalFormat::RGBA8, msaa_);
+    framebufferms_->addColorAttachment(TextureInternalFormat::RGBA16, msaa_);
     framebufferms_->addDepthAttachment(TextureInternalFormat::Depth24Stencil8, msaa_);
     assert(framebufferms_->isComplete());
     framebuffer_ = Framebuffer::create(resolution_[0], resolution_[1]);
-    framebuffer_->addColorAttachment(TextureInternalFormat::RGBA8);
+    framebuffer_->addColorAttachment(TextureInternalFormat::RGBA16);
     framebuffer_->addDepthAttachment(TextureInternalFormat::Depth24Stencil8);
     assert(framebuffer_->isComplete());
     effect_framebuffer_ = Framebuffer::create(resolution_[0], resolution_[1]);
-    effect_framebuffer_->addColorAttachment(TextureInternalFormat::RGBA8);
+    effect_framebuffer_->addColorAttachment(TextureInternalFormat::RGBA16);
     assert(effect_framebuffer_->isComplete());
     renderer.initialize();
     checkGLError();
